// src/app/generated/commander/model/mediaFileType.ts
/**
 * plexhelp-commander API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type MediaFileType = 'MOVIE' | 'TV';

export const MediaFileType = {
    Movie: 'MOVIE' as MediaFileType,
    Tv: 'TV' as MediaFileType
};